==> 01_Data_Merge.ipynb <==
# 1. 필요한 라이브러리 불러오기

import pandas as pd

# 2. 데이터 불러오기

active = pd.read_excel('./origin_data/액티브 중소법인 재무보유.xlsx', sheet_name=0)  # 기업기요_외감 데이터만 불러오기
closed = pd.read_excel('./origin_data/휴폐업 중소법인 재무보유.xlsx', sheet_name=0)  # 기업개요 데이터만 불러오기
finance = pd.read_table('./origin_data/재무데이터.txt', sep='\t', encoding='cp949')
active_business_information = pd.read_excel('./origin_data/액티브 중소법인 재무보유.xlsx', sheet_name=2) # 액티브 중소법인 중 휴폐업 이력이 있는 기업들 불러오기
closed_business_information = pd.read_excel('./origin_data/휴폐업 중소법인 재무보유.xlsx', sheet_name=1) # 휴폐업 중소법인들의 휴폐업 이력 불러오기

## 2.1. 데이터 확인

### 2.1.1. 액티브 중소법인 기업개요_외감
- 행: 28982개
- 열: 28개

print(f"액티브 중소법인의 기업개요 데이터 row 개수: {active.shape[0]}개")
print(f"액티브 중소법인의 기업개요 데이터 feature 개수: {active.shape[1]}개")

### 2.1.2 액티브 중소법인 휴폐업 이력
- 행: 4827개
- 열: 5개

print(f"액티브 중소법인의 휴폐업 이력 데이터 row 개수: {active_business_information.shape[0]}개")
print(f"액티브 중소법인의 휴폐업 이력 데이터 feature 개수: {active_business_information.shape[1]}개")

### 2.1.3. 휴폐업 중소법인 기업개요
- 행: 6739개
- 열: 28개

print(f"휴폐업 중소법인의 기업개요 데이터 row 개수: {closed.shape[0]}개")
print(f"휴폐업 중소법인의 기업개요 데이터 feature 개수: {closed.shape[1]}개")

### 2.1.4. 휴폐업 중소법인 휴폐업 이력
- 행: 6840개
- 열: 5개

print(f"휴폐업 중소법인의 휴폐업 이력 데이터 row 개수: {closed_business_information.shape[0]}개")
print(f"휴폐업 중소법인의 휴폐업 이력 데이터 feature 개수: {closed_business_information.shape[1]}개")

### 2.1.5. 재무데이터
- 행: 109142개
- 열: 46개

print(f"재무 데이터 row 개수: {finance.shape[0]}개")
print(f"재무 데이터 feature 개수: {finance.shape[1]}개")

## 2.2. 데이터 확인 결과
- 액티브 중소법인의 경우 운영중인 기업들 중에서 휴폐업 이력이 있는 경우만 데이터가 수집되었기 때문에 휴폐업 이력이 있는 데이터가 더 적게 있다.
- 휴폐업 중소법인의 경우 기업개요 데이터 보다 휴폐업 이력이 더 많은데, 한 기업에서 휴폐업 이력이 여러 번 나타났을 것으로 추정된다.
    - 따라서 추후 확인이 필요할 것 같다.
- 재무 데이터는 아직까지는 큰 특징이 없었다.

### 2.2.1 휴폐업 데이터에서 휴폐업 이력이 여러 번 나타난 기업이 있는지 확인하기
- 휴폐업 데이터의 경우 기업개요와 휴폐업 이력의 개수가 맞지 않았다.
- 따라서 휴폐업 이력이 여러 번 나타났는지, 휴폐업 이력에는 있지만 기업개요에는 없는 것이 있는지 확인해본다.

# 기업개요 데이터와 휴폐업 이력 데이터의 사업자등록번호 비교하기
closed_biz_no = sorted(closed['BIZ_NO'].unique().tolist())
closed_business_information_biz_no = sorted(closed_business_information['BIZ_NO'].unique().tolist())

if closed_biz_no == closed_business_information_biz_no:
    print("기업개요 데이터와 휴폐업 이력이 가지고 있는 사업자등록번호가 같습니다.")
else:
    print("기업개요 데이터와 휴폐업 이력이 가지고 있는 사업자등록번호가 다릅니다.")

# 어느 것이 더 많은 사업자등록번호를 가지고 있는지 확인하기
print(f"기업개요 데이터의 사업자등록번호 개수: {len(closed_biz_no)}개")
print(f"휴폐업 이력 데이터의 사업자등록번호 개수: {len(closed_business_information_biz_no)}개")

# 휴폐업 이력 데이터에만 있는 사업자등록번호 확인하기
closed_set = set(closed_biz_no)
closed_business_information_set = set(closed_business_information_biz_no)

check_closed_biz_no = closed_business_information_set.difference(closed_set)

print("휴폐업 이력에만 있는 사업자등록번호: ", end='')
print(check_closed_biz_no)

### 2.2.2. 확인 결과
- 휴폐업 이력에만 있는 데이터는 7개다.
- 7개라 무시할 수도 있지만, 휴폐업을 예측해야하기 때문에 중요한 데이터로 사용될 여지가 있다.
- 따라서 액티브 중소법인에 해당 데이터가 있는지 확인이 필요하다.

### 2.2.3. 휴폐업 중소법인 데이터의 휴폐업 이력에만 있는 데이터가 액티브 중소법인에 있는지 확인하기


exist_closed_biz_no = list(check_closed_biz_no) # 휴폐업 중소법인의 휴폐업 이력에만 있는 사업자등록번호
active_biz_no = active['BIZ_NO'].unique().tolist() # 액티브 중소법인의 사업자등록번호

exist_closed_biz_no_set = set(exist_closed_biz_no)
active_biz_no_set = set(active_biz_no)

check_result = exist_closed_biz_no_set.difference(active_biz_no_set)

if check_result == check_closed_biz_no:
    print("액티브 중소법인에도 없습니다.")
else:
    print("액티브 중소법인에 있습니다.")

### 2.2.4. 확인 결과
- 액티브 중소법인에도 없는 데이터 이므로 사용할 수 없다.
- 따라서 추후 합칠 때, 휴폐업 중소법인의 기업개요를 기준으로 휴폐업 이력 데이터와 합치기로 한다.

# 3. 파악하기 쉽도록 데이터 값 변경 해주기

## 3.1. 액티브 중소법인과 휴폐업 중소법인의 기업개요 데이터 값 변경하기

def change_sectors(value):
    """
    업종코드를 업종명으로 변경하는 함수입니다.
    value(str)
    ---------------------------------------
    input = 업종코드
    retrun  = 업종명
    ---------------------------------------
    A: 신용금고
    B: 은행
    C: 카드
    F: 생명보험
    I: 손해보험
    M: 제조
    O: 기타금융
    S: 증권
    V: 창업투자회사
    """
    sectors = {'A': '신용금고', 'B': '은행', 'C': '카드', 'F': '생명보험', 'I': '손해보험', 
               'M': '제조', 'O': '기타금융', 'S': '증권', 'V': '창업투자회사'}
    return sectors[value]

def change_company_size(value):
    """
    기업규모를 한글로 변경하는 함수입니다.
    ---------------------------------------
    input = 기업규모(범주형 숫자)
    retrun  = 기업규모(한글)
    ---------------------------------------
    1: 대기업
    2: 중소기업
    3: 기타(비영리법인)
    4: 중견기업
    5: 대기업(금융)
    6: 중견기업
    """
    company_size = {1: '대기업', 2: '중소기업', 3: '기타(비영리법인)', 4: '중견기업', 5: '대기업(금융)', 6: '중견기업'}
    return company_size[value]

def change_is_public_enterprise(value):
    """
    공기업구분을 한글로 변경하는 함수입니다.
    ---------------------------------------
    input = 공기업구분(범주형 숫자)
    retrun  = 공기업구분(한글)
    ---------------------------------------
    1: 공기업
    2: 일반기업
    """
    public_enterprise = {1: '공기업', 2: '일반기업'}
    return public_enterprise[value]

def change_is_corporation(value):
    """
    법인개인구분을 한글로 변경하는 함수입니다.
    ---------------------------------------
    input = 법인개인구분(범주형 숫자)
    retrun  = 법인개인구분(한글)
    ---------------------------------------
    1: 법인
    2: 개인
    """
    corporation = {1: '법인', 2: '개인'}
    return corporation[value]

def change_is_head_office(value):
    """
    본점지점구분을 한글로 변경하는 함수입니다.
    ---------------------------------------
    input = 본점지점구분(범주형 숫자)
    retrun  = 본점지점구분(한글)
    ---------------------------------------
    1: 법인
    2: 개인
    """
    head_office = {1: '본점', 2: '지점'}
    return head_office[value]

def change_type_of_public_institution(value):
    """
    공공기관유형을 한글로 변경하는 함수입니다.
    ---------------------------------------
    input = 국외투자법인여부(범주형 숫자)
    retrun  = 국외투자법인여부(한글)
    ---------------------------------------
    0: 해당사항없음
    9: 정부기구
    2: 공기업/준정부기관
    3: 지자체공단/공사
    """
    type_of_public_institution = {0: '해당사항없음', 9: '정부기구', 2: '공기업/준정부기관', 3: '지자체공단/공사'}
    return type_of_public_institution[value]

def change_establishment(value):
    """
    설립구분을 한글로 변경하는 함수입니다.
    ---------------------------------------
    input = 설립구분(범주형 숫자)
    retrun  = 설립구분(한글)
    ---------------------------------------
    1: 주식
    2: 합자
    3: 합명
    4: 유한
    5: 조합
    6: 정부투자기관
    7: 개인
    8: 학교
    9: 병원
    10: 단체.협회
    11: 미정의
    99: 기타
    """
    establishment = {1: '주식', 2: '합자', 3: '합명', 4: '유한', 5: '조합', 6: '정부투자기관', 
                    7: '개인', 8: '학교', 9: '병원', 10: '단체/협회', 11: '미정의', 99: '기타'}
    return establishment[value]

active_change_values = active.copy()
closed_change_values = closed.copy()

active_change_values['BZ_TYP'] = active['BZ_TYP'].apply(change_sectors)
closed_change_values['BZ_TYP'] = closed['BZ_TYP'].apply(change_sectors)

active_change_values['CMP_SCL'] = active['CMP_SCL'].apply(change_company_size)
closed_change_values['CMP_SCL'] = closed['CMP_SCL'].apply(change_company_size)

active_change_values['PBCO_GB'] = active['PBCO_GB'].apply(change_is_public_enterprise)
closed_change_values['PBCO_GB'] = closed['PBCO_GB'].apply(change_is_public_enterprise)

active_change_values['PSN_CORP_GB'] = active['PSN_CORP_GB'].apply(change_is_corporation)
closed_change_values['PSN_CORP_GB'] = closed['PSN_CORP_GB'].apply(change_is_corporation)

active_change_values['HDOF_BR_GB'] = active['HDOF_BR_GB'].apply(change_is_head_office)
closed_change_values['HDOF_BR_GB'] = closed['HDOF_BR_GB'].apply(change_is_head_office)

active_change_values['PB_ORG_TYP'] = active['PB_ORG_TYP'].apply(change_type_of_public_institution)
closed_change_values['PB_ORG_TYP'] = closed['PB_ORG_TYP'].apply(change_type_of_public_institution)

active_change_values['ESTB_GB'] = active['ESTB_GB'].apply(change_establishment)
closed_change_values['ESTB_GB'] = closed['ESTB_GB'].apply(change_establishment)

## 3.2. 액티브와 중소법인과 휴폐업 중소법인의 휴폐업 이력 데이터의 값 변경하기

def change_status(value):
    """
    휴폐업상태를 한글로 변경해주는 함수입니다.
    ---------------------------------------
    input = 휴폐업상태(범주형 숫자)
    retrun  = 휴폐업상태(한글)
    ---------------------------------------
    1: 부도
    2: 휴업
    3: 폐업
    4: 피합병
    5: 휴업취소
    6: 폐업취소
    """
    status = {1: '부도', 2: '휴업', 3: '폐업', 4: '피합병', 5: '휴업취소', 6: '폐업취소'}
    return status[value]

active_business_information_change_values = active_business_information.copy()
closed_business_information_change_values = closed_business_information.copy()

active_business_information_change_values['CLSBZ_GB'] = active_business_information['CLSBZ_GB'].apply(change_status)
closed_business_information_change_values['CLSBZ_GB'] = closed_business_information['CLSBZ_GB'].apply(change_status)

active_business_information_change_values['CLSBZ_GB'].value_counts()

closed_business_information_change_values['CLSBZ_GB'].value_counts()

# 4. 액티브 중소법인과 휴폐업 중소법인의 기업개요, 휴폐업 이력 데이터 합치기

active_final = pd.merge(active_change_values, active_business_information_change_values, how='left', on='BIZ_NO')
closed_final = pd.merge(closed_change_values, closed_business_information_change_values, how='left', on='BIZ_NO')

data = pd.concat([active_final, closed_final], ignore_index=True)

data.head()

# column명 변경하기
df_column = pd.read_excel('./origin_data/넘블 챌린지 개방데이터 레이아웃.xlsx', sheet_name=0)

df_column = df_column[['영문칼럼명', '한글칼럼명']].iloc[:33]
df_column = df_column.drop_duplicates().reset_index(drop=True)

df_column

data_columns = data.columns.tolist()
origin_columns = df_column['영문칼럼명'].values
new_data_columns = []
for col_name in data_columns:
    if col_name in origin_columns:
        new_data_columns.append(df_column.loc[df_column['영문칼럼명'] == col_name]['한글칼럼명'].values[0])
    else:
        new_data_columns.append(col_name)

data.columns = new_data_columns

print(f"합친 후 데이터의 크기: {data.shape}")

# 5. 데이터 정합성 파악하기
- 휴폐업 데이터의 경우, 이력성으로 데이터가 쌓이고, 다양한 입수처로부터 정보를 수집하기 때문에 데이터의 정합성이 맞지 않는 경우가 있을 수 있다.
- 따라서 추후 모델링에 악영향을 끼칠 수 있기 때문에 신뢰성있는 가장 최신의 데이터만 사용한다.

# 휴폐업 이력이 있거나 (휴폐업구분의 값에 업이 들어간 경우(휴업, 폐업, 휴업 취소, 폐업 취소)),
# 종료일자가 99991231이 아닌 경우(최신 데이터가 아닌 경우)에 해당하는 row의 index 추출
drop_idx = data.loc[(data['휴폐업구분'].str.contains('업')) & (data['종료일자'] != 99991231.0)].index

# 추출한 index를 drop
df_drop = data.drop(drop_idx)

print(f"정합성이 맞지 않는 {data.shape[0] - df_drop.shape[0]}개의 데이터가 제거되었습니다.")
print(f"제거 후 데이터프레임의 크기 {df_drop.shape}")

# 액티브와 휴폐업 합치고 내보내기
df_drop.to_csv('./custom_data/active_closed_all.csv', encoding='cp949', index=False)

# 5. 재무 데이터와 비재무 데이터(기업 개요) 합치기
- 기업 당 사업자등록번호가 고유한 값이기 때문에 사업자등록번호를 기준으로 합친다.

finance.shape[0], df_drop.shape[0]

finance.columns

# 데이터를 합칠 key에 해당하는 feature name 변경
finance.rename(columns={'사업자번호':'사업자등록번호'}, inplace=True)

- 기업의 연도별 결산 데이터를 확인해야하기 때문에 재무 데이터를 기준으로 두고 기업 개요 데이터를 합친다.

financial_and_corporation = pd.merge(finance, df_drop, how='inner', on='사업자등록번호')

print(f"재무 데이터와 비재무 데이터를 합친 후 크기: {financial_and_corporation.shape}")

# 6. 내보내기

financial_and_corporation.to_csv('./custom_data/base_data.csv', encoding='cp949', index=False)
