==> 04_Top10_Corporation.ipynb <==
## 연도별 업종대분류를 기준으로 top10기업을 구하기
- top10 선정 기준은 총자산순이익률
    - 총자산이익률 = 당기순이익 / 총자산

# 1. 필요한 라이브러리 불러오기

import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

# 2. 데이터 불러오기

df = pd.read_csv('./custom_data/feature_engineering_idurstry_code.csv', encoding='cp949')

df.shape

df.columns

# 3. 총자산이익률 계산하기

## 3.1. 계산 전 특이사항

df.loc[df['사업자등록번호'] == 4168194212][['결산년월', '자산총계', '당기순이익(손실)', '기업명', '휴폐업구분', '상태발생일자']]

df.loc[df['사업자등록번호'] == 6498800195][['결산년월', '자산총계', '당기순이익(손실)', '기업명', '휴폐업구분', '상태발생일자']]

- 폐업한 기업의 경우 폐업 전에 청산의 이유로 자산이 심하게 줄고 당기순이익이 급상승하는 패턴을 보이는 기업이 있다.
- 따라서 추후 총자산순이익율의 상위 10개를 구할 때는 폐업하지 않은 기업으로 범위를 좁힌다.

## 3.2. 계산을 위해 새로운 데이터프레임 만들기

# 필요한 피처만 가져오기
df_split = df[['사업자등록번호', '결산년월', '자산총계', '당기순이익(손실)', '기업명', '업종대분류', '휴폐업구분']]

# 정수형으로 되어있는 결산년월에서 연도 추출하기
def pop_year(value):
    return str(value)[:4]

df_split['년도'] = df_split['결산년월'].apply(pop_year)

# 자산총계와 당기순이익의 결측치는 0으로 보간
df_split.fillna(0, inplace=True)

# 공식에 따른 총자산순이익률 계산
df_split['총자산순이익률'] = df_split['당기순이익(손실)'] / df_split['자산총계']

df_split.head()

df_split['업종대분류'].value_counts()

- 공공 행정, 국방 및 사회보장 행정과 농업, 임업 및 어업의 경우 개수가 너무 적다.
- top10을 뽑아낼 수도 없고, 적은 데이터의 평균을 구하는 것도 의미가 없다고 판단하여 둘의 평균총자산순이익률은 0으로 둔다.

# 평균총자산순이익률을 구할 때 공공 행정, 국방 및 사회보장 행정과 농업, 임업 및 어업 제외
indurstry_list = df_split['업종대분류'].unique().tolist()[:-2]

indurstry_list

# 자산총계가 0인 경우 계산하면 inf가 나오므로 0으로 변환
df_final = df_split.replace([np.inf, -np.inf], 0)

# 위에서 분석한 결과를 토대로 영업중인 기업에 대해서만 구함
df_final = df_final.loc[df_final['휴폐업구분'] == 0]

## 3.3. 연도별 업종대분류에 따른 총자산순이익률 top10 기업 구하고 평균 구하기

list_2018 = []
for indurstry in indurstry_list:    
    df_temp = df_final.loc[(df_final['년도'] == '2018') & (df_final['업종대분류'] == indurstry)].drop_duplicates(['기업명']).sort_values('총자산순이익률', ascending=False).head(10)
    mean_value = df_temp['총자산순이익률'].mean()
    list_2018.append(mean_value)
list_2018.append(0.0)
list_2018.append(0.0)

list_2019 = []
for indurstry in indurstry_list:
    df_temp = df_final.loc[(df_final['년도'] == '2019') & (df_final['업종대분류'] == indurstry)].drop_duplicates(['기업명']).sort_values('총자산순이익률', ascending=False).head(10)
    mean_value = df_temp['총자산순이익률'].mean()
    list_2019.append(mean_value)
list_2019.append(0.0)
list_2019.append(0.0)

list_2020 = []
for indurstry in indurstry_list:
    df_temp = df_final.loc[(df_final['년도'] == '2020') & (df_final['업종대분류'] == indurstry)].drop_duplicates(['기업명']).sort_values('총자산순이익률', ascending=False).head(10)
    mean_value = df_temp['총자산순이익률'].mean()
    list_2020.append(mean_value)
list_2020.append(0.0)
list_2020.append(0.0)

list_2021 = []
for indurstry in indurstry_list:
    df_temp = df_final.loc[(df_final['년도'] == '2021') & (df_final['업종대분류'] == indurstry)].drop_duplicates(['기업명']).sort_values('총자산순이익률', ascending=False).head(10)
    mean_value = df_temp['총자산순이익률'].mean()
    list_2021.append(mean_value)
list_2021.append(0.0)
list_2021.append(0.0)

# 4. 내보내기

# 아까 제외한 업종대분류 추가를 위해 다시 정의
concat_indurstry_list = df_split['업종대분류'].unique().tolist()

# 데이터프레임으로 만들기
concat_list = []
for mean_list, year in zip([list_2018, list_2019, list_2020, list_2021], ['2018', '2019', '2020', '2021']):
    df_temp = pd.DataFrame({'년도': [year]*18,
                            '업종대분류':concat_indurstry_list,
                            '평균총자산이익률':mean_list})
    concat_list.append(df_temp)
final = pd.concat(concat_list)

final

final.to_csv('mean_top10_per_year.csv', encoding='cp949', index=False)
