==> 07_Analysis_Infromation_Gain.ipynb <==
# 1. 데이터 불러오기

import pandas as pd
import numpy as np

## 1.1. 2018년 미포함 + 윈저라이징 완료 데이터

import json

json_path = './custom_data/finance_information_gain_new.json'

with open(json_path, 'r') as j:
    contents=json.loads(j.read())

df = pd.Series(contents)
df = df.to_frame()
df = df.sort_values(by=0, ascending=False)
df = df.reset_index()
df.columns = ['변수명', 'information_gain']

## 1.2. 2018년 포함 + 윈저라이징 미완 데이터

import json

json_path = './custom_data/finance_information_gain.json'

with open(json_path, 'r') as j:
    contents=json.loads(j.read())

df_new = pd.Series(contents)
df_new = df_new.to_frame()
df_new = df_new.sort_values(by=0, ascending=False)
df_new = df_new.reset_index()
df_new.columns = ['변수명', 'information_gain']

## 1.3. 합쳐서 비교
- 상위 10개 중 변경된 변수
    1. 총자산영업이익률 : -
    2. 총자산순이익률 : -
    3. 자기자본영업이익률 : -
    4. 자기자본비율 : ▲ 1
    5. 유보율 : ▲ 4
    6. 당좌비율 : -
    7. 자기자본순이익률 : ▼ 3
    8. 유동비율 : ▼ 1
    9. 사내유보대자기자본비율 : ▼ 1
    10. 순운전자본회전율 : -
<br><br>

- 상위 10개 중 빠진 변수는 없음. 10개 내에서 변동이 이루어짐

1. 성장성
    - 당기순이익증가율 : 19
    - 영업이익증가율 : 20
    - 자기자본증가율 : 21
    - 총자본증가율 : 22
    - 매출액증가율 : 24
<br><br>
2. 수익성
    - 총자산영업이익률 : 1
    - 총자산순이익률 : 2
    - 자기자본영업이익률 : 3
    - 자기자본순이익률 : 7
    - 사내유보 대 자기자본비율 : 9
    - 매출액경상이익률 : 16
    - 매출액순이익률 : 17
    - 매출액영업이익률 : 18
    - 매출총이익률 : 25
    - 수지비율 : 29
    - 금융비용부담률 : 30
<br><br>
3. 활동성
    - 순운전자본회전율 : 10
    - 타인자본회전율 : 11
    - 총자본회전율 : 12
    - 자기자본회전율 : 13
    - 당좌자산회전율 : 14
    - 유동자산회전율 : 15
    - 운전자본회전율 : 23
    - 재고자산회전율 : 28
<br><br>
4. 생산성
    - 유보율 : 5
<br><br>
5. 안정성
    - 자기자본비율 : 5
    - 당좌비율 : 6
    - 유동비율 : 8
    - 재고자산 대 순운전자본비율 : 26
    - 매출채권 대 매입채무비율 : 27

- 각 항목별로 상위 순위 변수 1~3개 선정
- 사용할 변수 (11개)
    1. 성장성 : 당기순이익증가율, 영업이익증가율
    2. 수익성 : 총자산영업이익률, 총자산순이익률, 자기자본영업이익률
    3. 활동성 : 순운전자본회전율, 타인자본회전율
    4. 생산성 : 유보율
    5. 안정성 : 자기자본비율, 당좌비율, 유동비율

pd.concat([df, df_new], axis=1)

# 2. 사용할 변수 추출

use_vars = ['당기순이익증가율', '영업이익증가율', '총자산영업이익률', '총자산순이익률', '자기자본영업이익률',
            '순운전자본회전율', '타인자본회전율', '유보율', '자기자본비율', '당좌비율',
            '유동비율']

raw_data = pd.read_csv('./custom_data/all_finance_data.csv', encoding='cp949')
raw_data.columns

use_data = raw_data.iloc[:, 0:2]
use_data

for var in use_vars:
    for i in raw_data.columns:
        if var in i[:len(var)]:
            use_data[i] = raw_data[i]

print(len(use_data.columns))
print(use_data.columns)

# 3. 내보내기

use_data.to_csv('./custom_data/use_finance_data.csv', encoding='cp949')
