==> 09_Feature_Engineeting_Flagship_Indurstry.ipynb <==
# 1. 피처엔지니어링 3 : 지역특화산업 여부 파악하기

1990년말~2000년 초반부터 우리나라에서는 지역 균형 발전을 명분으로 수도권(서울, 경기, 인천)을 제외한 지방 기업을 대상으로 지역특화산업 육성사업을 별도의 예산으로 지원하고 있으며 창업기업 및 스타트업들이 많은 지원을 받을 수 있는 제도이다.

특정 지역의 지역특화산업에 해당하는 기업이라면, 혜택을 받을 가능성이 높아지고 이는 휴폐업의 여부와 연관이 있을 것이라고 예상한다. 따라서 우리는 연도별 지역특화산업을 조사하고, 해당하는 기업과 해당하지 않는 기업을 나누는 파생변수를 생성하려 한다.

# 2. 데이터 수집
지역특화산업은 KIAT에서 관리하다가 2022년부터 중소벤처기업부에서 관리한다. 우리가 필요한 데이터는 2019~2021년의 데이터이므로 KIAT 홈페이지에서 연도별 각 지역의 지역특화산업 공고문을 다운받아 수기로 옮기는 작업을 통해 지역특화산업의 핵심산업코드를 얻을 수 있었다.

수집한 결과 2019~2021년의 지역특화산업과 핵심산업코드는 아래와 같다.

#### 2019년 지역주력산업 산업코드
- 강원도
    - 웰니스식품산업
        - 핵심코드: 10797, 10211, 10212, 10501, 10730, 11111
    - 세라믹복합신소재산업
        - 핵심코드: 20129, 24113, 30399, 23222, 26299, 20412, 23211
    - 레저휴양지식서비스산업
        - 핵심코드: 58221, 58222, 62010, 62021, 62090, 63120
- 경상남도
    - 지능형기계산업
        - 핵심코드: 28111, 28119, 29162, 29229, 29280
    - 항공산업
        - 핵심코드: 22241, 26224, 31311, 31312, 31322
    - 나노융합부품산업
        - 핵심코드: 20129, 22192, 23211, 26129, 26299
    - 항노화바이오산업
        - 10797, 20423, 21101, 27191, 27192
- 경상북도
    - 지능형디지털기기산업
        - 핵심코드: 26422, 26211, 27211, 27112, 28121, 27199
    - 하이테크성형가공산업
        - 핵심코드: 20202, 29294, 24191, 24121, 24199, 24123
    - 바이오뷰티산업
        - 핵심코드: 10211, 10797, 10302, 20423, 10742, 20499
    - 기능성섬유산업
        - 핵심코드: 13104, 13213, 13219, 13999, 23995, 29261
- 광주광역시
    - 스마트가전산업
        - 핵심코드: 27215, 28121, 28122, 28511, 28519, 29172
    - 복합금형산업
        - 핵심코드: 22241, 22249, 22251, 22259, 25914, 25913, 26294
    - 디지털생체의료산업
        - 핵심코드: 27191, 27199, 27192, 27122
    - 광융합산업
        - 핵심코드: 26111, 26112, 26221, 26222, 26223, 26410, 27193, 27301, 28301
    - 지역연고산업- 치기공산업, 덴탈헬스케어, 광기술응용 의료기기
        - 핵심코드: 27192, 27191, 27199, 27112, 27329 / 27191, 27199, 27122, 27192, 20422 / 27199, 27301, 27309, 28909, 27219
- 대구광역시
    - 의료헬스케어산업
        - 핵심코드: 27112, 27191, 27192, 27199, 27213
    - 분산형 에너지산업
        - 핵심코드: 26222, 28111, 28123, 58221, 61210
    - 첨단소재부품산업
        - 핵심코드: 13213, 22192, 25911, 29294, 25913
    - 지역연고 산업 - 한방건강식품, 침장제품
        - 핵심코드: 10301, 10302, 10619, 10711, 10741, 10742, 13221, 13222, 13223, 13910
- 대전광역시
    - 무선통신융합산업
        - 핵심코드: 26295, 26111, 31311 28121, 26410, 58221, 62021
    - 로봇지능화산업
        - 핵심코드: 58221, 58222, 28909, 26121, 27216, 62010, 29280
    - 바이오기능성소재산업
        - 핵심코드: 21220, 21102, 20119, 10797, 10795
    - 지역연고산업 - 고기능성산업용포장재산업, 센글라스렌즈산업
        - 핵심코드: 22212, 22231, 22232, 22250, 22291, 22299 / 27119, 27321, 27329, 73209
- 부산광역시
    - 바이오메디컬산업
        - 핵심코드: 10796, 10797, 21300, 27192, 27199
    - 클린에너지산업
        - 핵심코드: 25122, 26221, 26299, 28111, 28112
    - 지능정보서비스산업
        - 핵심코드: 27215, 58211, 58221, 62010, 31312
    - 지능형기계부품산업
        - 핵심코드: 27213, 28121, 28122, 29120, 29229
- 세종특별자치도
    - 첨단수송기기부품산업
        - 핵심코드: 26295, 30310, 30320, 30332, 30399
    - 정밀의료산업
        - 핵심코드: 20423, 21210, 21300, 26299, 27199
    - 지역연고산업
        - 핵심코드: 10799, 10801, 10802, 10301, 10302, 10619, 10711, 10741, 10742
- 울산광역시
    - 친환경자동차부품산업
        - 핵심코드: 30121, 30310, 30320, 30331, 30399
    - 조선해양산업
        - 핵심코드: 27216, 29176, 31111, 31113, 31114
    - 첨단화학신소재산업
        - 핵심코드: 20119, 20132, 20202, 20203, 20499
    - 친환경에너지산업
        - 핵심코드: 20121, 25122, 25123, 28111, 28112, 28202
    - 지역연고산업 - 금형산업, 나노코팅소재산업
        - 핵심코드: 24321, 29294 / 25921, 25923
- 전라남도
    - 바이오헬스케어소재
        - 핵심코드: 10742, 10797, 20423, 21102, 21300
    - 에너지신산업
        - 핵심코드: 28111, 28119, 28123, 28909, 35114
    - 첨단운송기기부품
        - 핵심코드: 25122, 25999, 29299, 30322, 58222
    - 청색/청정환경
        - 핵심코드: 23211, 29172, 29173, 29175, 27216
- 전라북도
    - 농생명소재식품산업
        - 핵심코드: 10309, 10797, 10801, 21220, 58221
    - 지능형기계부품산업
        - 핵심코드: 29210, 29241, 29294, 29299, 62090
    - 탄소복합소재산업
        - 핵심코드: 23995, 20203, 13101, 20501, 17110
    - 해양설비기자재산업
        - 핵심코드: 31113, 20202, 25113, 28119, 29176
- 제주특별자치시
    - 청정헬스푸드산업
        - 핵심코드: 10129, 10212, 10302, 10309, 10796
    - 지능형관광디지털콘텐츠산업
        - 핵심코드: 58211, 58222, 62010, 73203, 90221
    - 스마트그리드산업
        - 핵심코드: 34019, 34020, 58221, 72122, 74100
    - 지역연고산업 - 기능성 식품, 관광서비스콘텐츠
        - 핵심코드: 10122, 10212, 10302, 10309, 10796, 11111, 11119, 58211, 58222, 62010, 63120, 63991
- 충청남도
    - 차세대디스플레이산업
        - 핵심코드: 26211, 26212, 29272, 26221, 26223
    - 친환경자동차부품산업
        - 핵심코드: 30320, 30331, 30393, 30332, 28202
    - 바이오식품산업
        - 핵심코드: 10742, 10796, 10797, 10795, 10802
    - 지역연고산업 - 바이오식품산업, 생활용품산업
        - 핵심코드: 10799, 10801, 10802 / 10799, 20421, 21101, 21300
- 충청북도
    - 스마트IT부품산업
        - 핵심코드: 26111, 26222, 29271, 26129, 27199
    - 바이오헬스산업
        - 핵심코드: 21101, 21210, 21300, 27192, 21102, 21220, 10797, 27191
    - 수송기계소재부품산업
        - 핵심코드: 25119, 30320, 30332, 30391, 25941, 30331, 30400, 30392

* * *

#### 2020년 지역주력산업 산업코드
- 강원도
    - 웰니스식품산업 
        - 핵심코드: 10211, 10212, 10501, 10730, 10797, 11111
    - 세라믹복합신소재산업
        - 핵심코드: 20129, 20412, 23211, 23222, 24113, 26299, 30399
    - 레저휴양지식서비스산업
        - 핵심코드: 58221, 58222, 62010, 62021, 62090, 63991
- 경상남도
    - 지능형기계산업
        - 핵심코드: 28111, 28119, 29162, 29229, 29280
    - 항공산업
        - 핵심코드: 22241, 26224, 31311, 31321, 31322
    - 나노융합부품산업
        - 핵심코드: 20131, 20499, 26299, 29294
    - 항노화바이오산업
        - 핵심코드: 10797, 20423, 27191, 27192
- 경상북도
    - 지능형디지털기기산업
        - 핵심코드: 26422, 27211, 28121, 28202, 26211, 27112, 27119, 26291
    - 바이오뷰티산업
        - 핵심코드: 10211, 10302, 10742, 10797, 20423, 20499
    - 하이테크성형가공
        - 핵심코드: 20202, 24191, 24199, 24132, 29294, 24121, 24123, 25913
    - 기능성섬유산업
        - 핵심코드: 13104, 13213, 13219, 13999, 23995, 29261
- 광주광역시
    - 디지털생체의료산업
        - 핵심코드: 27191, 27192, 27193, 27199
    - 복합금형산업
        - 핵심코드: 22241, 22249, 22251, 22259, 25914, 25913, 29294, 25921, 25923
    - 스마트가전산업
        - 핵심코드: 27215, 28121, 28122, 28511, 28519, 29172
    - 광융합산업
        - 핵심코드: 26111, 26112, 26221, 26222, 26410, 27301, 28301, 28421, 28422
- 대구광역시
    - 의료헬스케어산업
        - 핵심코드: 27112, 27191, 27192, 27199, 27213
    - 분산형에너지산업
        - 핵심코드: 26222, 28111, 28123, 58221, 61210
    - 첨단소재부품산업
        - 핵심코드: 22192, 23999, 25912, 22299, 25921
    - 협력산업육성 - 전기자율차, 첨단신소재
        - 핵심코드: 30391, 30392, 30320, 30331, 30332, 30399 / 29223, 29229, 30320, 30331, 30332, 30399
- 대전광역시
    - 무선통신융합산업
        - 핵심코드: 26129, 26429, 58222, 26295, 58221
    - 로봇지능화산업
        - 핵심코드: 27215, 27219, 29294, 58222, 62010
    - 바이오기능성소재산업
        - 핵심코드: 10797, 20423, 21102, 70113, 13999, 27213
    - 지역연고산업 - 에너지, 워터 그리드 통합기기 및 시스템 개발
        - 핵심코드: 20119, 27215, 28119, 28123, 28202, 29175, 62021, 62090, 72122
- 부산광역시
    - 지능정보서비스산업
        - 핵심코드: 27215, 31312, 58211, 58212, 58222, 59112, 59120, 62010, 62021
    - 클린에너지산업
        - 핵심코드: 25122, 26221, 26291, 26299, 28111, 28112, 28114, 29119, 29131, 29132
    - 바이오메디컬산업
        - 핵심코드: 10796, 10797, 20423, 20499, 21210, 21220, 21300, 27192, 27199, 70111
    - 지능형기게부품산업
        - 핵심코드: 25929, 27213, 28121, 28123, 29120, 29199, 29229, 29280, 31322
- 세종특별자치도
    - 첨단수송기기부품
        - 핵심코드: 22241, 30310, 30320, 30332, 30339
    - 정밀의료산업
        - 핵심코드: 20423, 21210, 21300, 26299, 27199
- 울산광역시
    - 친환경자동차부품산업
        - 핵심코드: 30121, 30310, 30320, 30331, 30391, 30392, 30399
    - 조선해양산업
        - 핵심코드: 27216, 29176, 31111, 31113, 31114    
    - 첨단화학신소재산업
        - 핵심코드: 20119, 20132, 20202, 20203, 20499
    - 친환경에너지산업
        - 핵심코드: 20121, 25122, 25123, 28111, 28202
- 전라남도
    - 바이오헬스케어소재산업
        - 핵심코드: 10742, 10797, 20423, 21102, 21300
    - 에너지신산업
        - 핵심코드: 28111, 28123, 28114, 28121, 58221
    - 첨단운송기기산업
        - 핵심코드: 25122, 25999, 29299, 30332, 58222
    - 청색청정환경산업
        - 핵심코드: 23211, 29172, 29173, 29175, 27216
- 전라북도
    - 지능형기계부품산업
        - 핵심코드: 29210, 29299, 29241, 62090, 29294
    - 농생명식품산업
        - 핵심코드: 10309, 21220, 10797, 58221, 10801
    - 탄소복합소재산업
        - 핵심코드: 23995, 20501, 20203, 17110, 13101
    - 해양설비기자재
        - 핵심코드: 31113, 28119, 20202, 29176, 25113
- 제주특별자치시
    - 청정헬스푸드산업
        - 핵심코드: 10121, 10212, 10309, 10122, 10713, 10792, 10797, 10801, 11201, 11209
    - 지능형관광콘텐츠
        - 핵심코드: 58211, 58222, 62010, 73203, 90221, 58212, 62090, 63120, 63991, 75992
    - 스마트그리드산업
        - 핵심코드: 34019, 34020, 35119, 35130, 70119, 72121, 72122 ,72129, 75320 ,38210
- 충청남도
    - 바이오식품산업
        - 핵심코드: 10742, 10795, 10796, 10802, 10797
    - 차세대디스플레이산업
        - 핵심코드: 26211, 26221, 26212, 26223, 29272
    - 친환경자동차부품산업
        - 핵심코드: 30320, 30331, 30393, 30332, 28202
- 충청북도
    - 스마트IT부품산업
        - 핵심코드: 26111, 26129, 26222, 27199, 26271, 26211
    - 바이오헬스산업
        - 핵심코드: 21101, 21102, 21210, 21220, 10749
    - 수송기계소재부품산업
        - 핵심코드: 25119, 25941, 30320, 30331, 30332, 24222, 29294

* * *

#### 2021년 지역주력산업 산업코드
- 강원도
    - 천연물바이오소재산업
        - 핵심코드: 10796, 10797, 20423, 21102, 21210, 21300
    - 세라믹복합신소재산업
        - 핵심코드: 20129, 20499, 23222, 23993, 23999, 24113, 26299, 30399
    - ICT융합헬스산업
        - 핵심코드: 21300, 27112, 27199, 58221, 58222, 62010
- 경상남도
    - 스마트기계산업
        - 핵심코드: 26429, 27211, 28111, 28119, 29162, 29229, 29280, 30331
    - 첨단항공산업
        - 핵심코드: 31322, 31311, 31321, 31312, 22241, 23995, 58222
    - 나노융합스마트부품산업
        - 핵심코드: 20131, 20499, 22292, 26299, 26129, 24329
    - 항노화메디컬산업
        - 핵심코드: 10797, 20423, 21300, 27192, 27199, 27112
- 경상북도
    - 지능형디지털기기산업
        - 핵심코드: 26211, 26291, 26295, 26422, 27112, 27199, 28119, 28121, 28202, 30332
    - 첨단신소재부품가공산업
        - 핵심코드: 20202, 24121, 24199, 25912, 25913, 25914, 25921, 29294
    - 친환경융합섬유소재산업
        - 핵심코드: 13104, 13213, 13219, 13409, 13992, 13994, 13999, 23994, 29261
    - 라이프케어뷰티산업
        - 핵심코드: 10211, 10302, 10742, 20423, 20499, 21102, 21210, 21220, 21300
- 광주광역시
    - 지능형가전산업
        - 핵심코드: 28511, 28519, 29172
    - 광융합산업
        - 핵심코드: 26410, 27301, 28301, 28421, 28422
    - 스마트금형산업
        - 핵심코드: 22251, 22259, 25913, 25914, 29294
    - 디지털생체의료산업
        - 핵심코드: 27191, 27192, 27193, 27199
- 대구광역시
    - 디지털의료헬스케어산업
        - 핵심코드: 20423, 21220, 27112, 27191, 27192, 27199, 27212, 27213, 27215
    - 고효율에너지시스템산업
        - 핵심코드: 26222, 26223, 27216, 28111, 28122, 28123, 28202, 28302, 28909, 29271
    - 수송기기기계소재부품산업
        - 핵심코드: 13213, 22192, 22299, 25912, 25921, 29133, 29210, 29294, 29299, 30320
- 대전광역시
    - 차세대무선통신융합산업
        - 핵심코드: 26129, 26429, 58222, 26295, 58221
    - 지능형로봇산업
        - 핵심코드: 27215, 27219, 29294, 58222, 62010
    - 바이오메디컬산업
        - 핵심코드: 10797, 20423, 21101, 21102, 21210, 21300, 27213, 70113, 13999
- 부산광역시
    - 지능정보서비스산업
        - 핵심코드: 26429, 27215, 29222, 58211, 58212, 58221, 58222, 62010, 62021, 63991
    - 친환경미래에너지산업
        - 핵심코드: 25122, 26221, 26299, 28111, 28112, 28114, 29119, 29176, 29132, 29150
    - 라이프케어산업
        - 핵심코드: 10796, 10797, 20423, 20499, 21210, 21220, 21300, 27112, 27192, 27199
    - 첨단융합기게부품산업
        - 핵심코드: 25929, 27213, 28121, 28123, 29120, 29199, 29229, 29280, 30310
- 세종특별자치도
    - 스마트시티산업
        - 핵심코드: 28111, 28123, 28302, 61220, 62021, 63991, 74100, 75320
    - 스마트그린융합부품산업
        - 핵심코드: 20423, 21210, 21300, 22241, 26299, 27199, 30310, 30320, 30332, 30399
- 울산광역시
    - 그린모빌리티산업
        - 핵심코드: 30121, 30201, 30203, 30320, 30391, 30392, 30399
    - 스마트조선산업
        - 핵심코드: 26429, 27211, 27216, 31111, 31113, 31114
    - 미래화학신소재산업
        - 핵심코드: 20111, 20119, 20202, 20499, 22241, 22259, 24290, 27213, 29222
    - 저탄소에너지산업
        - 핵심코드: 20121, 20129, 25114, 25122, 25123, 28111, 28114, 28123, 28202, 28302
- 전라남도
    - 저탄소지능형소재부품산업
        - 핵심코드: 20202, 24199, 25119, 22251, 30399, 20499
    - 그린에너지산업
        - 핵심코드: 28111, 28114, 28121, 28123, 35114, 58221
    - 첨단운송기기부품산업
        - 핵심코드: 25122, 25999, 29299, 30332, 58222
    - 바이오헬스케어산업
        - 핵심코드: 10742, 10797, 20423, 21102, 21300
- 전라북도
    - 스마트농생명식품산업
        - 핵심코드: 10309, 10797, 10801, 10619, 20423, 21220, 58221, 10129, 10743, 21230
    - 미래지능형기계산업
        - 핵심코드: 29210, 29241, 29294, 58222, 28909, 29229, 62090, 30339, 25112, 26229
    - 탄소복합소재산업
        - 핵심코드: 23995, 20203, 13101, 13219, 20501, 17110, 26129, 13213, 19229
    - 조선해양에너지산업
        - 핵심코드: 31113, 20202, 25113, 28123, 25119, 28119, 29176, 28111, 35119, 35114
- 제주특별자치시
    - 스마트관광산업
        - 핵심코드: 58221, 58222, 59112, 62010, 63120, 70209, 73301, 75290, 75992, 91210
    - 청정바이오산업
        - 핵심코드: 10121, 10122, 10309, 10713, 10792, 10796, 10797, 11209, 20423, 70112
    - 그린에너지산업
        - 핵심코드: 28213, 34019, 34020, 35114, 35130, 61299, 72122, 72129, 74100
- 충청남도
    - 친환경모빌리티산업
        - 핵심코드: 30320, 28111, 28202, 30332, 26111, 26295
    - 스마트휴먼바이오
        - 핵심코드: 10797, 10802, 20423, 10795, 27112, 27199
    - 차세대디스플레이산업
        - 핵심코드: 26212, 26223, 26299, 29272, 26219, 22212
- 충청북도
    - 지능형IT부품산업
        - 핵심코드: 20121, 26111, 26112, 26129, 26211, 26222, 27211, 27216, 29133, 29271
    - 바이오헬스산업
        - 핵심코드: 20423, 21101, 21102, 21210, 21220, 21300, 27199
    - 수송기계소재부품산업
        - 핵심코드: 25119, 28020, 30201, 30310, 30320, 30331, 30332, 30391, 30392, 30399

우리는 이 데이터들을 json 형태로 저장하여 처리하였다.

# 3. 지역특화산업의 타당성 확인

지역특화산업이라고 무조건 폐업에 유리하진 않을 것이다. 따라서 우리는 연도별 지역특화산업 여부에 대한 파생변수를 이진 명목형 변수로 만들고, 동일하게 이진 명목형 변수인 휴폐업 구분과 연관성이 있는지 알아보기 위해 Two Sample Chi-Test를 진행해본다.

## 3-1. 라이브러리 불러오기

import pandas as pd
from scipy import stats
import json
import warnings

warnings.filterwarnings('ignore')

## 3-2. 데이터 불러오기

df = pd.read_csv('./custom_data/finance_address_indurstry.csv')

df.shape

## 3-3. 카이제곱검정을 실시하기 위한 형태로 데이터 변환

def change_target(value):
    """
    타겟값을 변환하기 위한 함수입니다.
    --------------------------------
    input = 휴폐업구분
    return = 폐업 or 휴업
    --------------------------------
    """
    if value in ['폐업', '휴업']:
        return '폐업'
    return '영업'

df['휴폐업구분'] = df['휴폐업구분'].apply(change_target)

def extract_year(value):
    """
    결산년월에서 년도만 추출하는 함수입니다.
    -------------------------------------
    input = 결산년월
    return = 년도
    """
    return str(value)[:4]
df['년도'] = df['결산년월'].apply(extract_year)

# 산업코드를 정수로 변환
df['산업코드1'] = df['산업코드1'].astype(int)

# 필요한 비재무 피처만 추출
df_split = df[['사업자등록번호', '년도', '시도', '시군구', '산업코드1', '업종중분류', '업종대분류', '휴폐업구분']]

df_split.head()

# 시군구 중 오타, 없어진 지역이 있어 올바른 값으로 변환

missing_address = {
    '천산이':'천안시',
    '부천':'부천시',
    '아산군':'아산시',
    '청원군':'청주시'
}

def change_sigungoo(value):
    """
    시군구 컬럼에 잘못된 값들이 있어 올바른 값으로 변환해주는 함수입니다.
    ---------------------------------------------------------------
    input = 시군구
    return = 올바른 값으로 변환된 시군구, 원래 올바른 시군구
    ---------------------------------------------------------------
    """
    if value in missing_address.keys():
        return missing_address[value]
    return value

df['시군구'] = df['시군구'].apply(change_sigungoo)

## 3-4. 문제점 1) 수도권의 특화산업
- 지역특화산업 육성사업의 시작 계기가 수도권보다 비수도권의 발전이 더뎌 시행된 것이므로 수도권은 특화산업이 없다.
- 따라서 우리는 수도권에 가장 많이 분포된 업종 중 n개를 특화산업으로 지정하여 해당 여부를 따진다.
- 수도권은 비수도권보다 많은 발전을 한 것으로 보이므로 비수도권의 특화산업 개수가 가장 적은 것과 동일한 개수로 n을 지정한다.
    - 만약 강원도의 특화산업이 19개로 가장 적었다면, 수도권은 업종의 분포를 내림차순 기준으로 19개를 추출하여 수도권의 특화산업으로 지정한다는 의미

# 비수도권의 지역특화산업 산업코드 불러오기
with open ("./custom_data/2019_flagship_indurstry.json", "r", encoding='UTF8') as f:
    flagship_indurstry_code_2019 = json.load(f)

with open ("./custom_data/2020_flagship_indurstry.json", "r", encoding='UTF8') as f:
    flagship_indurstry_code_2020 = json.load(f)

with open ("./custom_data/2021_flagship_indurstry.json", "r", encoding='UTF8') as f:
    flagship_indurstry_code_2021 = json.load(f)

capital_area_flagship_indurstry_code_2019 = {}
capital_area_flagship_indurstry_code_2020 = {}
capital_area_flagship_indurstry_code_2021 = {}

for year, flagship_indurstry_code, captial_area_dict in zip(['2019', '2020', '2021'], [flagship_indurstry_code_2019, flagship_indurstry_code_2020, flagship_indurstry_code_2021], 
                                                             [capital_area_flagship_indurstry_code_2019, capital_area_flagship_indurstry_code_2020, capital_area_flagship_indurstry_code_2021]):
    length = []
    for _, val in flagship_indurstry_code.items():
        length.append(len(val))
    min_indurstry_count = min(length)
    print(f"{year}년 비수도권의 지역특화산업 산업코드 최솟값: {min_indurstry_count}개")
    df_year = df.loc[df['년도'] == year]
    print(f"{year}년 서울특별시의 업종중분류 종류: {df_year[df['시도'] == '서울특별시']['업종중분류'].nunique()}개")
    captial_area_dict['서울특별시'] = df_year[df['시도'] == '서울특별시']['업종중분류'].value_counts()[:min_indurstry_count].index

    print(f"{year}년 인천광역시의 업종중분류 종류: {df_year[df['시도'] == '인천광역시']['업종중분류'].nunique()}개")
    captial_area_dict['인천광역시'] = df_year[df['시도'] == '인천광역시']['업종중분류'].value_counts()[:min_indurstry_count].index

    print(f"{year}년 경기도의 업종중분류 종류: {df_year[df['시도'] == '경기도']['업종중분류'].nunique()}개")
    captial_area_dict['경기도'] = df_year[df['시도'] == '경기도']['업종중분류'].value_counts()[:min_indurstry_count].index
    
    print("===================================================================")

# 수도권과 비수도권 딕셔너리 합치기
flagship_indurstry_code_2019.update(capital_area_flagship_indurstry_code_2019)
flagship_indurstry_code_2020.update(capital_area_flagship_indurstry_code_2020)
flagship_indurstry_code_2021.update(capital_area_flagship_indurstry_code_2021)

# 반복 작업을 위한 딕셔너리 생성
capital_and_non_capital_total_dict = {
    '2019':flagship_indurstry_code_2019,
    '2020':flagship_indurstry_code_2020,
    '2021':flagship_indurstry_code_2021
}

def centeral_indurstry(df):
    """
    지역특화산업 파생변수를 생성하는 함수입니다.
    ---------------------------------------------------------
    input = 년도, 시도, 업종중분류, 산업코드1이 있는 데이터프레임
    return = 지역특화산업에 해당하는 경우 1, 아닌 경우 0
    """
    year = df['년도']
    sido = df['시도']
    if sido in ['서울특별시', '인천광역시', '경기도']:
        if df['업종중분류'] in capital_and_non_capital_total_dict[year][sido]:
            return 1
        return 0
    else:
        if df['산업코드1'] in capital_and_non_capital_total_dict[year][sido]:
            return 1
        else:
            return 0

# 지역특화산업 파생변수 생성           
df['지역특화산업'] = df.apply(lambda x: centeral_indurstry(x), axis=1)

## 3-5. 년도별 지역특화산업 여부와 휴폐업 여부의 연관성 파악하기: Two Sample Chi-Test

# 년도별로 파악하기 위한 데이터프레임 나누기

df_2019 = df.loc[df['년도'] == '2019']
df_2020 = df.loc[df['년도'] == '2020']
df_2021 = df.loc[df['년도'] == '2021']

# 검정을 위한 교차표 생성

data_2019 = pd.crosstab(df_2019['지역특화산업'], df_2019['휴폐업구분'])
data_2020 = pd.crosstab(df_2020['지역특화산업'], df_2020['휴폐업구분'])
data_2021 = pd.crosstab(df_2021['지역특화산업'], df_2021['휴폐업구분'])

def calculate_chi_test(data):
    """
    Two Sample Chi-Test를 수행하는 함수입니다.
    -----------------------------------------
    input = 지역특화산업 여부와 휴폐업 여부의 교차표
    return = 교차표에 카이제곱검정으로 산출한 p-value
    -----------------------------------------
    """
    _, p_value, _, _ = stats.chi2_contingency(data, correction=False)
    return p_value

for idx, data in enumerate([data_2019, data_2020, data_2021]):
    p_value = calculate_chi_test(data)
    if p_value > 0.05:
        print(f"{idx+2019}년의 지역특화산업 여부는 p-value가 {round(p_value, 3)}으로 0.05보다 크기 때문에 휴폐업 여부와 연관이 없습니다.")
    else:
        print(f"{idx+2019}년의 지역특화산업 여부는 p-value가 {round(p_value, 3)}으로 0.05보다 작기 때문에 휴폐업 여부와 연관이 있습니다.")

df_2019 = df_2019[['사업자등록번호', '결산년월', '지역특화산업']]
df_2020 = df_2020[['사업자등록번호', '결산년월', '지역특화산업']]
df_2021 = df_2021[['사업자등록번호', '결산년월', '지역특화산업']]

df_2019.to_csv('./flagship_indurstry_2019.csv', index=False, encoding='cp949')
df_2020.to_csv('./flagship_indurstry_2020.csv', index=False, encoding='cp949')
df_2021.to_csv('./flagship_indurstry_2021.csv', index=False, encoding='cp949')

## 3-5. 전체 데이터에서 지역특화산업 여부와 휴폐업 여부의 연관성 파악하기: Two Sample Chi-Test

data_total = pd.crosstab(df['지역특화산업'], df['휴폐업구분'])
p_value_total = calculate_chi_test(data_total)

if p_value_total > 0.05:
        print(f"전체 데이터에서 지역특화산업 여부는 p-value가 {round(p_value_total, 3)}으로 0.05보다 크기 때문에 휴폐업 여부와 연관이 없습니다.")
else:
    print(f"전체 데이터에서 지역특화산업 여부는 p-value가 {round(p_value_total, 3)}으로 0.05보다 작기 때문에 휴폐업 여부와 연관이 있습니다.")

# 4. 결론
지역특화산업 여부는 휴폐업 여부와 년도별로 관련이 있지만, 종합적으로는 관련이 없다. 따라서 현 모델에서는 사용하지 않는다.
