==> 10_Feature_Engineering_Finance_Grade.ipynb <==
# KSK지표

- 기업의 성장률이 휴폐업에 영향을 끼친다고 판단하였다.
- 새롭게 파생 변수들을 기반으로 지표를 만들고자 한다.

# 필요한 라이브러리 불러오기

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
sns.set(font_scale=0.8)
plt.rcParams['font.family'] = 'Malgun Gothic'

# 1. 데이터 불러오기

main = pd.read_csv('./custom_data/finance_address_indurstry.csv')
df  = pd.read_csv('./custom_data/active_closed_all.csv', encoding='cp949')
magic = pd.merge(main, df[['사업자등록번호', '휴폐업구분']], how='left', on='사업자등록번호')
industry = magic[['사업자등록번호','결산년월','업종대분류','업종중분류','시도','시군구']]
target = magic[['휴폐업구분']]

# 2. 새롭게 생성한 변수


- 총자본 증가율
- 영업이익 증가율
- 당기순이익증가율
- 자기자본증가율
- 매출액 증가율
- 매출총이익률
- 매출영업이익률
- 매출액경상이익률
- 매출액순이익률
- 총자산영업이익률
- 자기자본영엉이익률
- 금융비용부담률
- 수지비율
- 사내비율 대 자기자본비율
- 총자산순이익률
- 총자산회전율
- 자기자본회전율
- 타인자본회전율
- 유동자산회전율
- 재고자산회전율
- 당좌자산회전율
- 순운전자본회전율
- 운전자본회전율
- 유보율
- 자기자본비율
- 유동비율
- 당좌비율
- 재고자산 대 순운전자본비율
- 매출채권 대 매입채무비율


## 2.1 성장성

- 당기순이익증가율
- 영업이익증가율
- 자기자본증가율
- 총자본증가율
- 매출액증가율  

## 2.2. 수익성

- 총자산영업이익률
- 총자산순이익률
- 자기자본영업이익률
- 자기자본순이익률
- 사내유보대자기자본비율
- 매출액경상이익률
- 매출액순이익률
- 매출액영업이익률
- 매출액총이익률
- 수지비율
- 금융비율부담율

## 2.3. 활동성

- 순운전자본회전율
- 타인자본회전율
- 총자본회전율
- 자기자본회전율
- 당좌자산회전율
- 유동자산회전율
- 운전자본회전율
- 재고자산회전율

## 2.4. 생산성

- 유보율

## 2.5. 안정성

- 자기자본비율
- 당좌비율     
- 유동비율
- 재고자산대순운전자본비율
- 매출채권대매입채무비율


# 3. KSG 지표

## 3.1. 26개의 컬럼을 사용하여 만든 KSG지표
- 평균치의 상위 40 프로 이상인 컬럼이 24개 이상이면 SSS
- 평균치의 상위 40 프로 이상인 컬럼이 22개 이상이면 SS
- 평균치의 상위 40 프로 이상인 컬럼이 20개 이상이면 S
- 평균치의 상위 40 프로 이상인 컬럼이 18개 이상이면 AAA
- 평균치의 상위 40 프로 이상인 컬럼이 16개 이상이면 AA
- 평균치의 상위 40 프로 이상인 컬럼이 14개 이상이면 A
- 평균치의 상위 40 프로 이상인 컬럼이 11개 이상이면 B
- 평균치의 상위 40 프로 이상인 컬럼이 9개 이상이면 C
- 평균치의 상위 40 프로 이상인 컬럼이 7개 이상이면 D
- 평균치의 상위 40 프로 이상인 컬럼이 5개 이상이면 E
- 평균치의 상위 40 프로 이상인 컬럼이 5개 이하면 F

## 3.2. 사용할 피처들 저장

Total = magic[['당기순이익증가율','영업이익증가율','자기자본증가율','총자본증가율','매출액증가율','총자산영업이익률','총자산순이익률','자기자본영업이익률',
'자기자본순이익률','매출액경상이익률','매출액순이익률','매출액영업이익률','매출총이익률',
'순운전자본회전율','타인자본회전율','총자본회전율','자기자본회전율','당좌자산회전율','유동자산회전율','운전자본회전율','재고자산회전율',
'자기자본비율','당좌비율','유동비율','재고자산대순운전자본비율','매출채권대매입채무비율']]

## 3.3. 컬럼별 상위 40프로에 해당하는 기업 카운트

Total['count'] = 0
for column in Total.columns:
    quantile_60 = Total[column].quantile(0.5)
    Total.loc[Total[column] >= quantile_60, 'count'] += 1

## 3.4. 등급 나누기

for i in range(0,len(Total['count'])):

    if Total['count'].loc[i] >= 24 :
        Total['count'].loc[i] = 'SSS'
    elif Total['count'].loc[i] >= 22 :
        Total['count'].loc[i] = 'SS'
    elif Total['count'].loc[i] >= 20 :
        Total['count'].loc[i] = 'S'
    elif Total['count'].loc[i] >= 18 :
        Total['count'].loc[i] = 'AAA'
    elif Total['count'].loc[i] >= 16 :
        Total['count'].loc[i] = 'AA'
    elif Total['count'].loc[i] >= 14 :
        Total['count'].loc[i] = 'A'
    elif Total['count'].loc[i] >= 11 :
        Total['count'].loc[i] = 'B'
    elif Total['count'].loc[i] >= 9 :
        Total['count'].loc[i] = 'C'
    elif Total['count'].loc[i] >= 7 :
        Total['count'].loc[i] = 'D'
    elif Total['count'].loc[i] >= 5 :
        Total['count'].loc[i] = 'E'
    else:
        Total['count'].loc[i] = 'F'

## 3.5. 타겟값과 등급 concat

Total = pd.concat([target,Total['count']],axis=1)

Total

als = pd.concat([industry,Total],axis=1)

count_list = ['F', 'E', 'D','C', 'B',  'A', 'AA','AAA', 'S','SS', 'SSS'  ]
temp_list=[]
for i in count_list:
    temp = len(als.query(f'count=="{i}"'))
    temp_list.append(temp)

# 4. 등급과 폐업의 비율 확인

def KSK(x):
    rank_list = ['SSS','SS','S','AAA','AA','A','B','C','D','E','F']
    list_1=[]
    list_2=[]
    for i in rank_list:
        try:
            temp1 = round((x.query(f'count == "{i}"')['휴폐업구분'].value_counts()[0]/len(x))*100,3)
            temp2 = round((x.query(f'count == "{i}"')['휴폐업구분'].value_counts()[0]/len(x.query(f'count == "{i}"')['휴폐업구분']))*100,3)
            list_1.append(temp1)
            list_2.append(temp2)
        except:
            print(f'{i}해당 등급은 없습니다')
    return list_1,list_2
Total_1,Total_2 = KSK(Total)
Total_df1 = pd.DataFrame(Total_1,index=['SSS','SS','S','AAA','AA','A','B','C','D','E','F'],columns=['전체'])
Total_df2 = pd.DataFrame(Total_2,index=['SSS','SS','S','AAA','AA','A','B','C','D','E','F'],columns=['전체'])

## 4.1. 각 등급의 폐업 수 / 전체 기업의 비율

Total_df1['전체'].to_list()

Total_df1

## 4.2. 각 등급의 폐업 수 / 해당 등급의 기업 수

Total_df2['전체'].to_list()

Total_df2

## 4.3. 확인결과
- F ~ SSS 등급으로 갈수록 더 낮은 폐업률을 보였기 때문에
- 타겟값 (휴,폐업)을 예측에 도움이 될것이라 판단한다.
