==> 11_Feature_Engineering_Industrial_Accident.ipynb <==
# 1. 기본 데이터프레임 만들기

import pandas as pd
import numpy as np

df = pd.read_csv('./finance_address_indurstry.csv')
df1 = df[['사업자등록번호', '결산년월', '업종대분류', '주소', '시도', '시군구', '휴폐업구분']]
df1

df2 = pd.read_csv('C:/study/niceDNB/numble/custom_data/active_closed_all.csv', encoding='cp949')
df2 = df2[['사업자등록번호', '기업명']]

industrial_accident_yn = pd.merge(df1, df2, how='left')
industrial_accident_yn['시군구'].fillna('세종시', inplace=True)
industrial_accident_yn

industrial_accident_yn['결산년월'] = industrial_accident_yn['결산년월'].astype('str')
industrial_accident_yn['년도'] = industrial_accident_yn['결산년월'].apply(lambda x: x[:4])
industrial_accident_yn

industrial_accident_yn['지역'] = industrial_accident_yn['시도'].map({'서울특별시':'서울', '경기도':'경기', '충청북도':'충북','인천광역시':'인천', '대구광역시':'대구', '부산광역시':'부산', '강원도':'강원', '경상북도':'경북', '대전광역시':'대전', '전라남도':'전남', '전라북도':'전북', '경상남도':'경남', '제주특별자치도':'제주', '충청남도':'충남', '울산광역시':'울산', '광주광역시':'광주', '세종특별자치시':'세종'})

industrial_accident_yn[industrial_accident_yn.기업명.str.contains('주식회사')]

# 2. 산업재해 파일 클렌징

industrial_accident_2019 = pd.read_excel('./19년_산업재해_발생기업.xlsx', sheet_name = None)
industrial_accident_2019 = pd.concat(industrial_accident_2019, ignore_index=True)

industrial_accident_2020 = pd.read_excel('./20년_산업재해_발생기업.xlsx', sheet_name = None)
industrial_accident_2020 = pd.concat(industrial_accident_2020, ignore_index=True)

industrial_accident_2021 = pd.read_excel('./21년_산업재해_발생기업.xlsx', sheet_name = None)
industrial_accident_2021 = pd.concat(industrial_accident_2021, ignore_index=True)

industrial_accident_2019

## 2.1. try1 : re 모듈

import re

a = '㈜진성씨티건설(원청)\n㈜진성산업건설(하청) (마곡지구 업무용지 C11-2블럭 복합'
b = re.sub(r'\([^)]*\)', '', a)
b

name_list_temp = industrial_accident_2019['사업장명'].to_list()
address_list_2019 = industrial_accident_2019['소재지'].to_list()


name_list_2019 = []
for i in name_list_temp:
    name_list_2019.append(i.replace('\n', ' ').replace('㈜', '').replace('(주)', '').replace('주식회사', ''))
    
for i in range(len(name_list_2019)):
    name_list_2019[i] = re.sub(r'\([^)]*\)', '', name_list_2019[i])


industrial_accident_2019['사업장명'] = name_list_2019
industrial_accident_2019

# 3. 산업재해여부 판단
## 3.1. 2019년
- 기업명이 1글자, 2글자인 경우 산업재해명단과 비교하기가 힘듬 -> 제외시킴
- 기업명이 3글자, 4글자인 경우도 'in'으로 산업재해여부를 판단하면 부정확한 결과가 나옴 -> 글자수와 기업명이 같은 경우만 산업재해로 판단해야 함
- 따라서, 산업재해기업명단을 띄어쓰기 단위로 나누어 리스트로 만듬

name_list_temp = industrial_accident_2019['사업장명'].to_list()
address_list_2019 = industrial_accident_2019['소재지'].to_list()

name_list_2019 = []
for i in name_list_temp:
    name_list_2019.append(i.replace('\n', ' ').replace('㈜', '').replace('(주)', ' ').replace('주식회사', ''))
    
for i in range(len(name_list_2019)):
    name_list_2019[i] = re.sub(r'\([^)]*\)', '', name_list_2019[i]).replace(')', '')

temp = ''
for i in name_list_temp:
    temp += i.replace(')', '') + ' '
 
temp = temp.split(' ')
temp = [i for i in temp if i ]
temp

# index_name_dic = {}
# for n in range(77350):
#     if industrial_accident_yn.iloc[n]['년도'] == '2019' and len(industrial_accident_yn.iloc[n]['기업명']) >= 3: 
#         if industrial_accident_yn.iloc[n]['기업명'] in temp:
#             name_index = temp.index(industrial_accident_yn.iloc[n]['기업명'])
#             print(f"index : {n}  /  기업명 : {industrial_accident_yn.iloc[n]['기업명']}  /  산업재해리스트 : {temp[name_index]}")
#             index_name_dic[n] = 1

def check(x):
    if x['년도'] == '2019' and len(x['기업명']) >= 3: 
        if x['기업명'] in temp:
            return 1
        else:
            return 0
            
industrial_accident_yn['산업재해'] = industrial_accident_yn.apply(check, axis=1)

temp_2019 = industrial_accident_yn[(industrial_accident_yn['년도'] == '2019')]

temp_2019[temp_2019['산업재해'] == 1]

## 3.3 2020년

# 산업재해명단에 있는 기업명 정제
# 정확하게 일치하는지를 판단하기 위해 띄어쓰기 단위로 나누어 리스트로 만듬

name_list_temp = industrial_accident_2020['사업장명'].to_list()
address_list_2020 = industrial_accident_2020['소재지'].to_list()

name_list_2020 = []
for i in name_list_temp:
    name_list_2020.append(str(i).replace('\n', ' ').replace('㈜', '').replace('(주)', ' ').replace('주식회사', ''))
    
for i in range(len(name_list_2020)):
    name_list_2020[i] = re.sub(r'\([^)]*\)', '', name_list_2020[i]).replace(')', '')

temp = ''
for i in name_list_temp:
    temp += str(i).replace(')', '') + ' '
 
temp = temp.split(' ')
temp = [i for i in temp if i ]


def check(x):
    if x['년도'] == '2020' and len(x['기업명']) >= 3: 
        if x['기업명'] in temp:
            return 1
        else:
            return 0

industrial_accident_yn['산업재해'] = industrial_accident_yn.apply(check, axis=1)

temp_2020 = industrial_accident_yn[(industrial_accident_yn['년도'] == '2020')]

temp_2020[temp_2020['산업재해'] == 1]

## 3.4. 2021년

# 산업재해명단에 있는 기업명 정제
# 정확하게 일치하는지를 판단하기 위해 띄어쓰기 단위로 나누어 리스트로 만듬

name_list_temp = industrial_accident_2021['사업장명'].to_list()
address_list_2021 = industrial_accident_2021['소재지'].to_list()

name_list_2021 = []
for i in name_list_temp:
    name_list_2021.append(str(i).replace('\n', ' ').replace('㈜', '').replace('(주)', ' ').replace('주식회사', ''))
    
for i in range(len(name_list_2021)):
    name_list_2021[i] = re.sub(r'\([^)]*\)', '', name_list_2021[i]).replace(')', '')

temp = ''
for i in name_list_temp:
    temp += str(i).replace(')', '') + ' '
 
temp = temp.split(' ')
temp = [i for i in temp if i ]


def check(x):
    if x['년도'] == '2021' and len(x['기업명']) >= 3: 
        if x['기업명'] in temp:
            return 1
        else:
            return 0

industrial_accident_yn['산업재해'] = industrial_accident_yn.apply(check, axis=1)

temp_2021 = industrial_accident_yn[(industrial_accident_yn['년도'] == '2021')]

temp_2021[temp_2021['산업재해'] == 1]

## 3.5. 파일합치기

industrial_accident_yn = pd.concat([temp_2019, temp_2020, temp_2021]).sort_index()
industrial_accident_yn

industrial_accident_yn[(industrial_accident_yn.산업재해 == 1)]

len(industrial_accident_yn[(industrial_accident_yn.산업재해 == 1) & (industrial_accident_yn.휴폐업구분 == '폐업')])

industrial_accident_yn[industrial_accident_yn.휴폐업구분 == '휴업']

# 4. 통계검정

def change_target(value):
    """
    타겟값을 변환하기 위한 함수입니다.
    --------------------------------
    input = 휴폐업구분
    return = 폐업 or 휴업
    --------------------------------
    """
    if value in ['폐업', '휴업']:
        return '폐업'
    return '영업'

industrial_accident_yn['휴폐업구분'] = industrial_accident_yn['휴폐업구분'].apply(change_target)

def change_yn(value):
    """
    타겟값을 변환하기 위한 함수입니다.
    --------------------------------
    input = 산업재해여부
    return = 해당 or 미해당
    --------------------------------
    """
    if value == 1:
        return '해당'
    return '미해당'

industrial_accident_yn['산업재해'] = industrial_accident_yn['산업재해'].apply(change_yn)

# 년도별로 파악하기 위한 데이터프레임 나누기

df_2019 = industrial_accident_yn.loc[industrial_accident_yn['년도'] == '2019']
df_2020 = industrial_accident_yn.loc[industrial_accident_yn['년도'] == '2020']
df_2021 = industrial_accident_yn.loc[industrial_accident_yn['년도'] == '2021']

# 검정을 위한 교차표 생성

data_2019 = pd.crosstab(df_2019['산업재해'], df_2019['휴폐업구분'])
data_2020 = pd.crosstab(df_2020['산업재해'], df_2020['휴폐업구분'])
data_2021 = pd.crosstab(df_2021['산업재해'], df_2021['휴폐업구분'])

from scipy import stats


def calculate_chi_test(data):
    """
    Two Sample Chi-Test를 수행하는 함수입니다.
    -----------------------------------------
    input = 산업재해해당 여부와 휴폐업 여부의 교차표
    return = 교차표에 카이제곱검정으로 산출한 p-value
    -----------------------------------------
    """
    _, p_value, _, _ = stats.chi2_contingency(data, correction=False)
    return p_value

for idx, data in enumerate([data_2019, data_2020, data_2021]):
    p_value = calculate_chi_test(data)
    if p_value > 0.05:
        print(f"{idx+2019}년의 산업재해해당 여부는 p-value가 {round(p_value, 3)}으로 0.05보다 크기 때문에 휴폐업 여부와 연관이 없습니다.")
    else:
        print(f"{idx+2019}년의 산업재해해당 여부는 p-value가 {round(p_value, 3)}으로 0.05보다 작기 때문에 휴폐업 여부와 연관이 있습니다.")

data_total = pd.crosstab(industrial_accident_yn['산업재해'], industrial_accident_yn['휴폐업구분'])
p_value_total = calculate_chi_test(data_total)

if p_value_total > 0.05:
        print(f"전체 데이터에서 산업재해 여부는 p-value가 {round(p_value_total, 3)}으로 0.05보다 크기 때문에 휴폐업 여부와 연관이 없습니다.")
else:
    print(f"전체 데이터에서 산업재해 여부는 p-value가 {round(p_value_total, 3)}으로 0.05보다 작기 때문에 휴폐업 여부와 연관이 있습니다.")
