==> 13_Feature_Engineering_ScoreCard.ipynb <==
# 1. 재무/비재무 점수 생성

import pandas as pd

df_for_merge = pd.read_csv('./냄훈.csv', encoding='cp949')
df_finance_ratio = pd.read_csv('./custom_data/all_finance_data.csv', encoding='cp949')

# 재무비율 컬럼 추출
use_finance_columns = df_finance_ratio.columns[:32]

# 재무비율에 대한 스코어카드를 생성하므로 재무비율 데이터만 추출
finance_ratio = df_finance_ratio[use_finance_columns]

# 타겟값과 연결짓기 위해 데이터 합치기
df = pd.merge(df_for_merge[['사업자등록번호', '결산년월', '휴폐업구분']], finance_ratio, on=['사업자등록번호', '결산년월'], how='left')

# 스코어카드 생성을 위해 X, y로 나누기
X = df.iloc[:, 3:]
y = df['휴폐업구분']

## 1-1. 재무 스코어카드 생성

from optbinning import Scorecard, BinningProcess
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression

df_application_train, df_application_test, y_train, y_test = train_test_split(
X, y, test_size=0.2, random_state=42)

# 타겟값을 제외한 모든 데이터의 컬럼명
list_features = X.columns.values

# 범주형 변수 컬럼명만 따로 빼기
list_categorical = df_application_train.select_dtypes(include=['object', 'category']).columns.values

# 선택방법?
selection_criteria = {"iv": {"min": 0.005, 'max':0.5, "strategy": "highest"}}

# Instatiate BinningProcess
binning_process = BinningProcess(
    categorical_variables=list_categorical,
    variable_names=list_features,
    selection_criteria=selection_criteria,
)

logreg = LogisticRegression(C=3, max_iter=1000, random_state=161)

# 스케일 방법 정의
scaling_method = "min_max"
scaling_method_data = {"min": 0, "max": 1000}

# 스코어카드 객체 정의
scorecard = Scorecard(
    binning_process=binning_process,
    estimator=logreg,
    scaling_method=scaling_method,
    scaling_method_params=scaling_method_data,
    intercept_based=False,
    reverse_scorecard=True,
)

scorecard.fit(df_application_train, y_train)

finance_column_list = ['총자본증가율', '영업이익증가율', '당기순이익증가율', '자기자본증가율', '매출액증가율', '매출총이익률',
                       '매출액영업이익률', '매출액경상이익률', '매출액순이익률', '총자산영업이익률', '자기자본영업이익률',
                       '자기자본순이익률', '금융비용부담률', '수지비율', '사내유보대자기자본비율', '총자산순이익률', '총자본회전율',
                       '자기자본회전율', '타인자본회전율', '유동자산회전율', '재고자산회전율', '당좌자산회전율', '순운전자본회전율',
                       '운전자본회전율', '유보율', '자기자본비율', '유동비율', '당좌비율', '재고자산대순운전자본비율',
                       '매출채권대매입채무비율']
len(finance_column_list)

# 스코어카드 테이블 생성
scorecard_summary = scorecard.table(style="detailed").round(3)

## 1-1-1. 스코어카드 내보내기

scorecard_summary.to_csv('./scorecard_finance_ratio.csv', index=False, encoding='cp949')

scorecard_summary

### 1-1-2. 재무점수 계산하기

import numpy as np

def change_bin(value):
    if value in ['Special', 'Missing']:
        return np.nan
    value = value.replace('(', '')
    value = value.replace(')', '')
    value = value.replace(' ', '')
    value = value.replace('[', '')
    value = value.replace(']', '')
    a = value.split(',')
    temp = []
    for i in a:
        if i == '-inf':
            temp.append(-np.inf)
        else:
            temp.append(float(i))
    return temp

chage_bin_scorecard = scorecard_summary.copy()
chage_bin_scorecard['Bin'] = scorecard_summary['Bin'].apply(change_bin)

chage_bin_scorecard.dropna(inplace=True)

chage_bin_scorecard

# 유보율은 스코어 계산이 안됐으므로 제거
X.drop('유보율', axis=1, inplace=True)

X_add_score = X.copy()

for column in X.columns:
    result = []
    for value in X[column]:
        scorecard = chage_bin_scorecard.loc[chage_bin_scorecard['Variable'] == column]
        for range_val, score in zip(scorecard['Bin'], scorecard['Points']):
            if range_val[0] <= value < range_val[1]:
                result.append(score)
    column_name = column+'점수'
    X_add_score[column_name] = result

finance_ratio_score = X_add_score.iloc[:, 29:]

df['재무비율점수'] = finance_ratio_score.sum(axis=1)

### 1-1-4. 재무점수 데이터 내보내기

df[['사업자등록번호', '결산년월', '재무비율점수', '휴폐업구분']].to_csv('./finance_ratio_score.csv', index=False, encoding='cp949')

df_for_merge.columns[41:]

## 1-2. 비재무 스코어카드 생성

df_electric = pd.read_csv('./산업재해_전기요금.csv', index_col=0)

df_statistice_electric = pd.merge(df_for_merge, df_electric, on=['사업자등록번호', '결산년월'], how='left')

df_split_statistics_eletric = df_statistice_electric[df_statistice_electric.columns[41:]]

df_split_statistics_eletric

y = df_statistice_electric['휴폐업구분']

from optbinning import Scorecard, BinningProcess
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression

df_application_train, df_application_test, y_train, y_test = train_test_split(
df_split_statistics_eletric, y, test_size=0.2, random_state=42)

# 타겟값을 제외한 모든 데이터의 컬럼명
list_features = df_split_statistics_eletric.columns.values

# 범주형 변수 컬럼명만 따로 빼기
list_categorical = df_application_train.select_dtypes(include=['object', 'category']).columns.values

# 선택방법?
selection_criteria = {"iv": {"min": 0.005, 'max':0.5, "strategy": "highest"}}

# Instatiate BinningProcess
binning_process = BinningProcess(
    categorical_variables=list_categorical,
    variable_names=list_features,
    selection_criteria=selection_criteria,
)

logreg = LogisticRegression(C=3, max_iter=1000, random_state=161)

# Define scaling method and values
scaling_method = "min_max"
scaling_method_data = {"min": 0, "max": 1000}

# Instatiate and fit Scorecard
scorecard = Scorecard(
    binning_process=binning_process,
    estimator=logreg,
    scaling_method=scaling_method,
    scaling_method_params=scaling_method_data,
    intercept_based=False,
    reverse_scorecard=True,
)

scorecard.fit(df_application_train, y_train)

scorecard_summary = scorecard.table(style="detailed").round(3)

scorecard_summary['Variable'].unique()

scorecard_summary

chage_bin_scorecard = scorecard_summary.copy()
chage_bin_scorecard['Bin'] = scorecard_summary['Bin'].apply(change_bin)

### 1-2-1. 비재무점수 계산하기

chage_bin_scorecard.dropna(inplace=True)

score_columns = ['혼자사는사람', '취업자증감', '고용률', '경제활동참가률', '실업률', '광공업생산지수', '재정자립도',
       '특허', '가구별소득', '노인비율', '외국인비율', '인구수', '도시면적', '전기사용량',
       '전기요금_minmax']

X_add_score = df_statistice_electric.copy()

for column in score_columns:
    result = []
    for value in df_statistice_electric[column]:
        scorecard = chage_bin_scorecard.loc[chage_bin_scorecard['Variable'] == column]
        for range_val, score in zip(scorecard['Bin'], scorecard['Points']):
            if range_val[0] <= value < range_val[1]:
                result.append(score)
    column_name = column+'점수'
    X_add_score[column_name] = result

X_add_score.columns[58:]

df_statistice_electric_score = X_add_score[X_add_score.columns[58:]]

df_statistice_electric['비재무점수'] = df_statistice_electric_score.sum(axis=1)

### 1-2-2. 비재무점수 데이터 내보내기

df_statistice_electric[['사업자등록번호', '결산년월', '비재무점수', '휴폐업구분']].to_csv('./statistics_score.csv', index=False, encoding='cp949')
